# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API for creating and managing jobs in the Selfier service.",
        "title": "Selfier Job Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/jobs": {
            "get": {
                "description": "Retrieves a list of all jobs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get all jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_job_adapter.JobResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_job_adapter.GinError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new job. The structure of the 'payload' depends on the 'type'.\nFor 'deselfie', the payload must be `{\"image_url\": \"string\"}`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Create a new job",
                "parameters": [
                    {
                        "description": "Create Job Request",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_job_adapter.CreateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_job_adapter.JobResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or payload",
                        "schema": {
                            "$ref": "#/definitions/internal_job_adapter.GinError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_job_adapter.GinError"
                        }
                    }
                }
            }
        },
        "/jobs/{id}": {
            "get": {
                "description": "Retrieves a single job by its unique ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get a job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_job_adapter.JobResponse"
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/internal_job_adapter.GinError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a job by its unique ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Delete a job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_job_adapter.GinError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_job_adapter.CreateJobRequest": {
            "type": "object",
            "required": [
                "payload",
                "type"
            ],
            "properties": {
                "payload": {
                    "type": "string",
                    "example": "{\"image_url\":\"https://example.com/image.jpg\"}"
                },
                "type": {
                    "enum": [
                        "deselfie"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/selfier_internal_job_core.JobType"
                        }
                    ],
                    "example": "deselfie"
                }
            }
        },
        "internal_job_adapter.GinError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_job_adapter.JobResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "an error occurred"
                },
                "id": {
                    "type": "string",
                    "example": "job-12345"
                },
                "payload": {
                    "type": "string",
                    "example": "{\"image_url\":\"https://example.com/image.jpg\"}"
                },
                "result": {
                    "type": "string",
                    "example": "{\"processed_url\":\"https://example.com/result.jpg\"}"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/selfier_internal_job_core.JobStatus"
                        }
                    ],
                    "example": "pending"
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/selfier_internal_job_core.JobType"
                        }
                    ],
                    "example": "deselfie"
                }
            }
        },
        "selfier_internal_job_core.JobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusProcessing",
                "StatusCompleted",
                "StatusFailed"
            ]
        },
        "selfier_internal_job_core.JobType": {
            "type": "string",
            "enum": [
                "deselfie"
            ],
            "x-enum-varnames": [
                "JobTypeDeselfie"
            ]
        }
    }
}